+++ src/main/java/com/jpmc/midascore/entity/UserRecord.java
@@ -1,5 +1,8 @@
 package com.jpmc.midascore.entity;
 
+import java.math.BigDecimal;
+import java.text.DecimalFormat;
+
 import jakarta.persistence.*;
 
 @Entity
@@ -13,12 +16,15 @@ public class UserRecord {
     private String name;
 
     @Column(nullable = false)
-    private float balance;
+    private BigDecimal balance;
+    
 
     protected UserRecord() {
     }
 
-    public UserRecord(String name, float balance) {
+   
+    
+    public UserRecord(String name, BigDecimal balance) {
         this.name = name;
         this.balance = balance;
     }
@@ -36,11 +42,11 @@ public class UserRecord {
         return name;
     }
 
-    public float getBalance() {
+    public BigDecimal getBalance() {
         return balance;
     }
 
-    public void setBalance(float balance) {
+    public void setBalance(BigDecimal balance) {
         this.balance = balance;
     }
 }
